Move mz-repr types to mz-catalog-types and remove mz-repr dep from mz-catalog-protos
Move RoleId, NetworkPolicyId, CatalogItemId, GlobalId, RelationVersion,
AclMode, and MzAclItem from mz-repr to mz-catalog-types. The original
modules in mz-repr re-export these types for backwards compatibility and
retain supplementary impls (RustType proto, Columnation, FixedSizeCodec).

This eliminates the mz-repr dependency from mz-catalog-protos, which now
imports all catalog types from the lightweight mz-catalog-types crate.
