+++ src/catalog-types/Cargo.toml
@@ -10,6 +10,11 @@ publish = false
 workspace = true
 
 [dependencies]
+anyhow = "1.0.100"
+bitflags = "1.3.2"
+columnar = "0.11.1"
+mz-lowertest = { path = "../lowertest", default-features = false }
+mz-ore = { path = "../ore", default-features = false }
 proptest = { version = "1.9.0", default-features = false, features = ["std"] }
 proptest-derive = "0.7.0"
 serde = { version = "1.0.219", features = ["derive"] }
