+++ src/repr/src/lib.rs
@@ -55,6 +55,8 @@ pub use crate::relation::{
     ColumnDiff, ColumnIndex, ColumnName, KeyDiff, NotNullViolation, PropRelationDescDiff,
     ProtoColumnName, ProtoColumnType, ProtoRelationDesc, ProtoRelationType, RelationDesc,
     RelationDescBuilder, RelationDescDiff, RelationVersion, RelationVersionSelector,
+    ast_version_to_relation_version, relation_version_to_ast_version,
+    relation_version_to_schema_id,
     ReprColumnType, ReprRelationType, SqlColumnType, SqlRelationType, UNKNOWN_COLUMN_NAME,
     VersionedRelationDesc, arb_relation_desc_diff, arb_relation_desc_projection,
     arb_row_for_relation,
