+++ src/sql/src/names.rs
@@ -22,7 +22,7 @@ use mz_ore::cast::CastFrom;
 use mz_ore::str::StrExt;
 use mz_repr::network_policy_id::NetworkPolicyId;
 use mz_repr::role_id::RoleId;
-use mz_repr::{CatalogItemId, GlobalId, RelationVersion};
+use mz_repr::{CatalogItemId, GlobalId};
 use mz_repr::{ColumnName, RelationVersionSelector};
 use mz_sql_parser::ast::visit_mut::VisitMutNode;
 use mz_sql_parser::ast::{CreateContinualTaskStatement, Expr, RawNetworkPolicyName, Version};
@@ -495,7 +495,8 @@ impl AstDisplay for ResolvedItemName {
 
                 if *print_id {
                     if let RelationVersionSelector::Specific(version) = version {
-                        let version: Version = (*version).into();
+                        let version: Version =
+                            mz_repr::relation_version_to_ast_version(*version);
                         f.write_str(" VERSION ");
                         f.write_node(&version);
                     }
@@ -1452,7 +1453,7 @@ impl<'a> NameResolver<'a> {
             },
             // Note: Return the specific version if one is specified, even if the feature is off.
             Some(v) => {
-                let specified_version = RelationVersion::from(v);
+                let specified_version = mz_repr::ast_version_to_relation_version(v);
                 match item.latest_version() {
                     Some(latest) if latest >= specified_version => {
                         RelationVersionSelector::Specific(specified_version)
